+++ frontend/app/components/patients/patient-tabs.tsx
@@ -1,11 +1,11 @@
 'use client'
 
 import type { Patient } from '@api/models/Patient'
-import { AppointmentCard } from '@components/AppointmentCard'
-import { BasicInformation } from '@components/BasicInformation'
-import { InsuranceCard } from '@components/InsuranceCard'
-import { SleepStudyCard } from '@components/SleepStudyCard'
-import { TreatmentCard } from '@components/TreatmentCard'
+import { AppointmentCard } from '@components/appointments/AppointmentCard'
+import { BasicInformation } from '@components/patients/BasicInformation'
+import { InsuranceCard } from '@components/patients/InsuranceCard'
+import { SleepStudyCard } from '@components/patients/SleepStudyCard'
+import { TreatmentCard } from '@components/patients/TreatmentCard'
 import { Button } from '@components/ui/button'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@components/ui/tabs'
 import { format, parseISO } from 'date-fns'
