+++ frontend/app/patients/[id]/edit/page.tsx
@@ -4,7 +4,7 @@ import { getPatient } from '@actions/patient/get-patient-action'
 import { getPatientCustomFields } from '@actions/patient/get-patient-custom-fields-action'
 import { updatePatient } from '@actions/patient/update-patient-action'
 import type { Patient } from '@api/models/Patient'
-import { PatientForm } from '@components/PatientForm'
+import { PatientForm } from '@components/forms/PatientForm'
 import { Button } from '@components/ui/button'
 import { X } from 'lucide-react'
 import { useRouter } from 'next/navigation'
